makeotf. Suppress a lot of overly verbose messages.
In MakeOTF.py:
Add new -V option to print all warnings and notices. The following are now shown only with this option:
- the 'makeotfexe' command line
- note about writing options file
- the note about checking that any glyphs in processed layer of a UFO font are up to date
- the note that a source font is being converted to a temporary Type1 font for input to makeotfexe

In addition, the project options file is now written only if the option '-sp' is used.

In makeotfexe:
The note about subroutine stack depth being reduced was entirely removed.
Add new -V option to print all warnings and notices. The following are now shown only with this option:
- note about source font before each warning during parsing
- note about GDEF class  types conflicting for each glyph which is used in more than one GDEF class context in the features file. An example is a glyph being used as a base glyph in one replacement rule, and as a mark glyph in another.