[tx]. Update Mac OSX Xcode project  to omit unused files.
This project has *.txt files that used to be the source for the *.h in-line help files. That was years ago, and we don't even have the tool that built the *.h from the *.txt files anymore. The *.h files are now the master source for the in-line help text.