[fontPDF] Add 'pageIncludeTitle' option (#379)
This is needed for diff'ing the PDFs that are generated by the *plot tools. By default each page contains a header (a.k.a. Title) that includes a time stamp, among other info. Each PDF page is then compressed using ASCII85. By removing the page header it's possible to do a straight diff of two PDFs, even if their pages are compressed.

In case raw (i.e. uncompressed) pages are needed, use the setPageCompression() method of pdfgen.py